back_fastapi/pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
PyJWT = {extras = ["crypto"], version = "^2.6.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
